Fix validators for new base_rest API
In the initial implementation of rest_api, the schema validators
had to be returned by methods in the same service as the method, named
after the endpoint's method with a prefix: "_validator_<method>" or
"_validator_return_<method>".

As we have a lot of endpoints methods in some services, we extracted
the validator methods in dedicated components with
"base.shopfloor.validator" and "base.shopfloor.validator.return" usages,
and methods of the same name as the endpoint's method.

With the new API, endpoints are decorated with "@restapi.method" and the
validator is defined there. Example:

    @restapi.method(
        [(["/<int:id>/get", "/<int:id>"], "GET")],
        output_param=restapi.CerberusValidator("_get_partner_schema"),
        auth="public",
    )

For backward compatilibity, base_rest patches the methods not decorated
and sets the "input_param" and "output_param" to call the
"_validator_<method>" or "_validator_return_<method>":

https://github.com/OCA/rest-framework/blob/abd74cd7241d3b93054825cc3e41cb7b693c9000/base_rest/models/rest_service_registration.py#L240-L250

This change adds the extra layer that will call the validator components
when one of these methods is called.

A follow-up commit will replace the getattr by a better fix with a
change in base_rest, but meanwhile, this commit will unblock the OCA/wms
CI.