Allow to use the barcode on Package Storage Type at reception screen
While in the step 'Product Packaging' of the reception screen
of an incoming transfer, allow to scan a barcode for a package
storage type. If a package storage type is found, then fill it
automatically and set the product packaging also from it.

In the reception screen, the order in which the product packaging
and the package storage type has been reversed, because it is the
latter which is now used while scanning, so it comes first now
in the user interface.