[FIX] shopfloor: handle backorders
When moves are validated among all scenarios, we are validating related
picking the standard way (with 'action_done()') when:

1) the moves to process are equal to all assigned moves of the current picking
   (so we can have waiting moves in the picking, in that case a normal
   backorder will be created by 'action_done()'), in other terms it is
   triggered when we are validating the last moves related to one picking,

2) the moves to process are linked to unprocessed ancestor/source moves.
   This condition is required because we want to avoid some business logic
   bound to the `_create_backorder` standard method, like the backorder
   strategy configured through the `stock_picking_backorder_strategy`
   module.

In all other cases, we want to process the current moves in their own
picking (use of `extract_and_action_done()` method), letting the current
picking `confirmed/waiting` with the remaining qties.

Point 1) was already implemented, this commit implements the point 2).

Also, location content transfer scenario was not using the
'validate_moves()' method (already used by all scenarios to validate
moves), this commit is fixing that.