Keep previously selected dest. location on screen
In the reception screen, when we select a destination location
and click on Next, the location becomes empty and a message asks
to set a dest. location.

This is because on the button click, the inverse method of
current_move_line_location_dest_id writes on move_line.location_dest_id,
and the value of current_move_line_location_dest_id is read again from
the compute method. The compute method erases the field's value.
If a destination has been set previously on the field, keep it instead.
To know if a destination has been set, we check if it is different from
the move, as the default location equal to the move's one.

Additionally, we don't empty the field's value if the allowed locations
is equal to the selected location: this should cover the hypothetical
case where a stock.move would have a leaf location as destination, we
want to keep the same in the move line.