[IMP] shopfloor: Add standard messages for Checkout Packing Info
This PR adds the option to choose predefined messages for the
parameter "Checkout Packing Info", that is set on the res.partner
and reused in other places. For this, the original text field
shopfloor_packing_info has been transformed into a Many2one that
points to the new model shopfloor.packing.info. Thus, the field
in the res.partner has been renamed to shopfloor_packing_info_id,
that has caused some renames in other files that used it. These
new messages are archived instead of deleted, to keep track of
old messages associated to existing (or archived) res.partners.

The new predefined messages can be created directly in the partner
or in the menu Inventory > Settings > Shopfloor.

To migrate the data, the existing data has been migrated by assuming
that the first line is the name while all the content is for the text
of the new model.