[13.0] stock_reception_screen exec picking action_done
In the flow of the `stock_reception_screen` module, the `action_done` of
a stock.picking is never called. It is only done so on stock.move.

As the state of the stock.picking is computed it ends up with the
correct value, but any logic in its `action_done` function is never
executed.

This change fixes that.