stock_storage_type: prevent useless writes on stock.location
Complement previous commits: e89b2819 and 461a26e0
We want to prevent writing these computed fields when unnecessary,
for instance, on non internal locations, we should not update anything
if the value was already correct.

However, when the value has to be computed (internal location, ...),
we want to write the value if it didn't change, so concurrent
transactions are retried.

An example: if we have a quant of 20 in a location, and 2 concurrent
transaction each take a quant,

0. Initial value for "location_is_empty": False
1. Tx A: set a qty done on move line for taking 12
2. Tx B: set a qty done on move line for taking 8

None of the transactions see that the other took the rest, none of them
change the `location_is_empty` to True. But if they both try to write
`location_is_empty: False` on the location, one of them is rejected and
retry after the other is committed: this time the second one see it
takes the last 8 and change `location_is_empty` to True.