shopfloor: checkout, allows delivery packages only
A delivery package is a package with a delivery packaging defined
('packaging_id' field). It is a package different than internal ones
(which are stored in the warehouse with their own constraints like
height...).
Here we enforce the use of such delivery packages when we perform a
checkout, this means that we can not use internal packages to perform a
delivery.

* 'list_dest_package' method ('/select_dest_package') is now listing only
  delivery packages
* 'scan_dest_package' and 'set_dest_package' methods accept only a
  delivery package
* 'scan_package_action' method accept only a delivery package