shopfloor: imp cluster_picking._scan_line_by_location
When processing a cluster of pickings we are working on existing move lines.

When we scan a location and there are some lines already processed
we can work on the next one.

There's no need to block users to ask to scan specific package/lot/product
if the system can determine that there's a line that can be safely processed.

Hence, _scan_line_by_location now apply the check only on pending move
lines matching the type of operation we are working on.