[13.0][IMP] putaway abc location sorting
Before this change, the list of location returned by the module would be
sorted by abc type first and then by max height. This sets a higher
priority for the abc type than the max height paramter on selecting a
suitable location.

Resulting in a less than optimal usage of height storage space.

Changing the sorting to be done first by max height and then by abc type
sould improve on that.

Also improve performance in the sorting of max height.