[FIX] shopfloor: single pack transfer, handle backorders
When processing a package with the single pack transfer scenario which
belongs to a transfer having several other packages, the remaining ones
are not put in a backorder and as such can not be processed anymore.

This commit ensure to put the processed package/package level in its own
move which will be validated in its own transfer, keeping the remaining
packages to process in the current transfer.