[FIX] stock_storage_type: Fix putaway compution when a strategy is defined
When the strategy has to find a leaf sub-location, do not return the
location itself as defined on a storage type putaway sequence.

Example: sequence is defined to first look in WH/Stock/area1 and then in
WH/Stock/area2. On area1 the strategy is to find a leaf sub-location
(strategy = Ordered Children Location) under area1. If area1 has
temporarily no sub-location, area1 should not be returned as a valid
leaf location.

Also, when searching for a leaf location, exclude non active locations