[FIX] stock_storage_type: Fix putaway compution when a strategy is defined
When the strategy has to find a leaf location, do not return non-leaf
locations defined on a storage type putaway sequence.

Example: sequence is defined to first look in WH/Stock/area1 and then in
WH/Stock/area2. On area1 the strategy is to find a leaf location under
area1. If area1 currently has no valid leaf location for the pack storage
type, area1 should not be returned as a valid leaf location.