Fix reception_screen next pack button
When receiving a move in multiple step (packages), the reception screen
moves back to the main screen. When it should offer to enter the next
pack. This is the normal workflow in v13.

Without this change for picking with a single move the new stock move
is assigned to a new picking and the current picking in the reception
screen is done.