shopfloor: Location content transfer - Fix wrong thing scanned
If you don't scan the right package/product/lot, raise the
correct error to the operator that he scanned the wrong
package/product/lot instead of raising a "Barcode not found"

In the initial step where you have to scan a location:
* add a check that the location is valid for the menu's picking types
* raise a "Location empty" error message if the location is empty
  instead of a "no pack found in location" (the pack error is for
  single_pack_transfer)

I changed the "Location empty" from info to error. I checked the
scenario using this message and seems to me to always be an error and
not an info.