Shopfloor: Add no_prefill_qty option
Co-authored-by: Matthieu Méquignon <matthieu.mequignon@camptocamp.com>
Co-authored-by: JuMiSanAr <jumisan134@gmail.com>
Co-authored-by: Thierry Ducrest <thierry@ducrest.net>

Add new option in scenario zone picking, cluster picking, checkout
When a product is scanned, increase the done quantity by 1
When a product packaging is scanned, increase the done quantity by the
product packaging quantity

Fix quantity picker sorting when there is a packaging with quantity 1.
Ensure Unit is always shown as last.