Add 'stock_warehouse_flow' to configure routing flows
New feature allowing to re-qualify the operation type of a stock move at
confirmation. The main criteria to requalify a move is the carrier, as this
one often implies a dedicated routing (we do not have the same flow, nor the
same carrier, depending if we ship a pallet or a bunch of products for instance).

This job was previously done by configuring the stock_dynamic_routing
with a domain on the carrier, but the dynamic routing was applied on
reservation, while we could already qualify moves at confirmation.