Add sublocation scan on shopfloor delivery scenario
When scanning a product barcode or a lot the scenario will search
for all related move lines and pick the first one by default.

For a shopfloor menu configuration with locations containing sub locations
this is not ideal. Because a user working at a sub location only wants
move lines related to that sub location to be selected.

To help with this situation, a user can first scan his working location
which is kept as long as no other location is scanned.
The sub location set can also be reset by the click of a button.

Also when the search for a move line to work on returns multiple move lines
related to different locations and no location has been preliminary scanned.
The user will be prompted to scann a location first.