sf_mobile_base: add report issue action
When users have troubles w/ the app
having an easy way to report an issue
with proper technical info is very handy.

Here we go by using direct emails
so that if there's an issue w/ the service
you bypass it completely.

You can enable and customized it via shopfloor_app_info params.