shopfloor, delivery: process lines of one delivery operation
If the same bulk product was present in several delivery operations (of
the same picking type), the first call to 'scan_deliver' was working, but with
the side-effect to set the `qty_done` for the lines of ALL delivery
operations related to this product.

This fix will process only the lines from one delivery operation at a
time, leaving the other lines untouched so they can be processed by
subsequent calls to `scan_deliver`, without raising the
`"No product found among current transfers."` error anymore (product was
not found because of the `qty_done` set during the first scan).