shopfloor, delivery: rework 'Process as pre-packaged' option
As the tooltip of the option was describing, its main goal is to
validate the delivery order (and create a backorder if needed) as soon
as a goods is scanned. This could be a package, a product, a packaging
or a lot.

Also, in all cases, when we scan a lot we want to process 1 qty (like
for product scanning), not all the line.