shopfloor: fix split move lines
The move line must be linked to the new split move before calling
_action_confirm.

In case of a reception picking where the supplier location bypass the
reservation, action_assign is called by _action_config here
https://github.com/odoo/odoo/blob/14.0/addons/stock/models/stock_move.py#L1177
which causes the creation of an unwanted new move line here
https://github.com/odoo/odoo/blob/14.0/addons/stock/models/stock_move.py#L1324
This is why we must relink the existing move line to the new move before
calling _action_confirm.

The new move line was also causing the original picking to change state
to waiting.