shopfloor: checkout add show_oneline_package_content option
On the menu item you can decide to enable the option
'show_oneline_package_content'.
Once enabled, on the select line screen,
if the package contains only one line
the lot and the product name will be shown.

The use case it tries to solve:
you have multiple packs in front of you
and you want to be able to scan any of the packs
using the lot or the product.