[IMP] stock_available_to_promise_release: Remove unused fields
For some historical motivations, some informations from the partner are stored on stock.move(zip_code, city) and stock.picking (zip_code, city, state_id). We can imagines UC where these fields were used into the list views to filter pickings and/or moves to release according to one of these criteria. Nevertheless, this use of these fields is specific to some specific customers and is not justified by the way the addon works. Moreover, since these fields are defined as related on the res.partner table, any modification to one of these fields on your database where the module is installed will freeze it for a long time and most probably ends in a concurrent update exception since it will trigger an update on all the pickings and moves where this partner is referenced. Last but not least, if you install this addon on a large database with hundreds of thousands or millions of rows in the related tables, it will stuck your upgrade process for a very long time.
If these fields are required for your customer it's safer to add them into a specific addon and put the necessary precautions to avoid any of the above issues