shopfloor: zone picking handle complete mix package
When a package has different product in it, it will be displayed with
multiple move lines in the `select_line` step. And they will need to be
move one after an other.

For such package, this change allows to move such package in one step.
In the selection of a line only one line for such package will be
displayed.

The quantity picker will not be displayed.