[IMP] stock_storage_type: Take into account complex location views structure
If a view location has a strategy == 'none' and a sequence to be applied, the first call
to _get_putaway_strategy in Odoo core will return the first internal child (and no strategy
will be applied at all).

So, if the location is a view and has no putaway rules for that product, substitute the
child_internal_location_ids field value to self.