[16.0][ADD] - stock_release_channel_shipment_advice_date
This module enhances the shipment advice planning process by incorporating
additional settings fields into the release channel. These settings allow for
the computation of scheduled dates for shipment arrival and departure from the
dock. These dates will then be automatically set in the shipment advice when it
is created through the planner.