shopfloor: fix checkout scan delivery packaging
On the `select_line` endpoint, when scanning a delivery package to pack every
thing in one go, the `no_prefill_qty` option is not taken into account.

Now when this option is on and no quantitiy have been set the user will
be asked to do so.

Also the confirmation to pack everything will be asked only if no
quantities are set (whatever the state of `no_prefill_qty`)