sf: zp ensure user scan a product at line selection
On line selection for the zone picking scenario.

Without this change the user may move a product without ever
scanning the product barcode or lot.
Because after scanning scanning a packge if only one line is left
the next screen with the quantity picker is shown to the user, where
the quantity can be changed manually or by scanning a product.