[FIX] stock_picking_batch_creation: Unset user on picking
Unset user on picking on unset on the batch.

By default the user is written on the batch, if a value is provided, the method 'assign_batch_user' is called on the pickings. But if you unset the user, this change is not propagated to the pickings and they remains assigned to the user. We override the write method on the stock.picking.batch model to force the uset of the user on the pickings in such a case