stock_release_channel_batch_mode_commercial_partner: fix tests
Previous code was setting the `pickings.need_release` field to False
to change the cache value even if it was a computed field without inverse.

Still, this field wasn't used at all to detect transfers to release. To
do so it's better to update the moves of relevant transfers (which is
also a stored field).