shopfloor_delivery_shipment: Group package by sales
When working from a location after a dock and then a location.
There can be a long list of packages without hints about the one
belonging to a delivery already started.

To improve this, the packages are grouped by sales order, keep on top of
the list the sales order started but not yet all loaded.