[IMP] stock_release_channel_process_end_time: process_end_date as date_deadline
Add a new option on the route definition to allow the use of the release channel's process end date as date deadline on stock moves created when releasing deliveries. This is required to allow the merge of generated moves. By default pulled moves keep the date deadline of their destination moves. If different moves for the same product from different sale orders are grouped into the same delivery, you expect that the system will generate only one pick operation. It's only possible if the generated pulled moves have the same date deadline.