[FIX] shopfloor: When a move line is postponed, reset the qty_done
If there are several move lines for the same product,
e.g.: a product stored in several locations, and related to
the same stock move, when the first proposed line is postponed and when the
operator validates the second line, shopfloor will try to validate the stock move
and both lines. This is not correct as the qty_done field still contains the reserved
quantity.

So, when postponing a line, reset the qty_done field to 0.0 in order not to embed
that quantity in the validated picking.