stock_warehouse_flow: Improve flow matching
Before this commit, applying a flow on a move would prevent
trying to find a new flow for it, as it changes picking_type_id,
which is the first criterion for flow matching.
This commits backs up original picking_type_id so a new flow can be
searched for, if move values are modified after a flow is applied.