[IMP] shopfloor: Add context key to control package total quantity computation
the total quantity per package is not relevant or displayed in all scenarios where the parser is used.
currently, it's only used in the zone_picking scenario

this PR introduces a context key to conditionally disable the total quantity computation based on the use
case. this avoids unnecessary access to stock.quant when the information isn't required,
improving performance and preventing potential concurrency issues